Rewritten shell installer.
Signed-off-by: ubi de feo <me@ubidefeo.com>

The new installer supports multiple packages supplied as paths (local or full).
It recursively goes through every file and creates a paths list to then
create folders on the board and copy all the files.
It accepts --mpy and --no-reset to convert files with mpy-cross and skip the
reset step at the end of the script.

Requires:
- mpremote
- mpy-cross